+++ MeetingBar/Notifications.swift
@@ -30,17 +30,38 @@ func registerNotificationCategories() {
     let acceptAction = UNNotificationAction(identifier: "JOIN_ACTION",
                                             title: "Join",
                                             options: .foreground)
-
-    let eventCategory =
-        UNNotificationCategory(identifier: "EVENT",
-                               actions: [acceptAction],
-                               intentIdentifiers: [],
-                               hiddenPreviewsBodyPlaceholder: "",
-                               options: [.customDismissAction, .hiddenPreviewsShowTitle])
+    
+    let snoozeUntilStartTime = UNNotificationAction(identifier: "SNOOZE_UNTIL_START_TIME",
+                                                    title: "Snooze until start time",
+                                                    options: .foreground)
+    
+    let snooze5Min = UNNotificationAction(identifier: "SNOOZE_5_MIN",
+                                          title: "Snooze for 5 min",
+                                          options: .foreground)
+    
+    let snooze15Min = UNNotificationAction(identifier: "SNOOZE_15_MIN",
+                                           title: "Snooze for 15 min",
+                                           options: .foreground)
+    
+    let snooze30Min = UNNotificationAction(identifier: "SNOOZE_30_MIN",
+                                           title: "Snooze for 30 min",
+                                           options: .foreground)
+    
+    let eventCategory = UNNotificationCategory(identifier: "EVENT",
+                                               actions: [acceptAction, snooze5Min, snooze15Min, snooze30Min, snoozeUntilStartTime],
+                                               intentIdentifiers: [],
+                                               hiddenPreviewsBodyPlaceholder: "",
+                                               options: [.customDismissAction, .hiddenPreviewsShowTitle])
+    
+    let snoozeEventCategory = UNNotificationCategory(identifier: "SNOOZE_EVENT",
+                                               actions: [acceptAction, snooze5Min, snooze15Min, snooze30Min],
+                                               intentIdentifiers: [],
+                                               hiddenPreviewsBodyPlaceholder: "",
+                                               options: [.customDismissAction, .hiddenPreviewsShowTitle])
 
     let notificationCenter = UNUserNotificationCenter.current()
 
-    notificationCenter.setNotificationCategories([eventCategory])
+    notificationCenter.setNotificationCategories([eventCategory, snoozeEventCategory])
 
     notificationCenter.getNotificationCategories { categories in
         for category in categories {
@@ -158,6 +179,42 @@ func scheduleEventNotification(_ event: EKEvent) {
     content.userInfo = ["eventID": event.eventIdentifier!]
     content.threadIdentifier = "meetingbar"
 
+    let trigger = UNTimeIntervalNotificationTrigger(timeInterval: timeInterval, repeats: false)
+    let request = UNNotificationRequest(identifier: "NEXT_EVENT", content: content, trigger: trigger)
+    center.add(request) { error in
+        if let error = error {
+            NSLog("%@", "request \(request) could not be added because of error \(error)")
+        } else {
+            NSLog("%@", "request \(request) was added")
+        }
+    }
+}
+
+func snoozeEventNotification(_ event: EKEvent, _ interval: SnoozeEventNotificationTime) {
+    requestNotificationAuthorization() // By the apple best practices
+    removePendingNotificationRequests()
+    
+    let now = Date()
+    let center = UNUserNotificationCenter.current()
+    var timeInterval = Double(interval.rawValue)
+    let content = UNMutableNotificationContent()
+
+    if Defaults[.hideMeetingTitle] {
+        content.title = "general_meeting".loco()
+    } else {
+        content.title = event.title
+    }
+    
+    if interval == .untilStart {
+        timeInterval = event.startDate.timeIntervalSince(now)
+    }
+    
+    content.categoryIdentifier = "SNOOZE_EVENT"
+    content.sound = UNNotificationSound.default
+    content.userInfo = ["eventID": event.eventIdentifier!]
+    content.threadIdentifier = "meetingbar"
+    content.body = "notifications_event_started_body".loco()
+    
     let trigger = UNTimeIntervalNotificationTrigger(timeInterval: timeInterval, repeats: false)
     let request = UNNotificationRequest(identifier: "NEXT_EVENT", content: content, trigger: trigger)
     center.add(request) { error in